bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
discord-pretty-help==1.3.4
discord.py @ git+https://github.com/Rapptz/discord.py@334ef1d7facce9dbfba2a6924bf57fc59bc827b5
dnspython==2.2.0
emojis==0.6.0
idna==3.3
multidict==6.0.2
psutil==5.9.0
pycparser==2.21
pymongo==4.0.1
PyNaCl==1.4.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
typing_extensions==4.1.1
yarl==1.7.2
